ceph: add custom zone endpoint for object-store multisite
If Ceph Zone and ZoneGroup endpoints lists are
modified manually after deployment they are modified
to include the original internal K8s clusterIPs if
the rook operator is restarted.

This commit allows for a custom endpoint to be
added for the Zone endpoints list. This custom
endpoint replaces the internal K8s clusterIPs
in the zone endpoints list so that the list
does not need to be modified manually anymore.

Signed-off-by: Ali Maredia <amaredia@redhat.com>
Signed-off-by: Jiffin Tony Thottan <thottanjiffin@gmail.com>